Update environments_names.tsv
testing Jen's understanding of fixing two-term strings: "shrubland" and variants were mapped to one acceptable term, ENVO:01000176, and one inappropriate term, ENVO:00000300- so the latter mappings were removed. "rivers" and variants were already deduplicated, but mapped to ENVO:00000022, which is not in use, so that entire mapping was replaced with the term that is in use, ENVO:01000253.